+++ src/map/party.c
@@ -736,9 +736,9 @@ static int party_optionchanged(int party_id, int account_id, int exp, int item, 
 		return 0;
 
 	//Flag&0x1: Exp change denied. Flag&0x10: Item change denied.
-	if(!(flag&0x01) && p->party.exp != exp)
+	if(!(flag&0x01) && p->party.exp != (unsigned int)exp)
 		p->party.exp=exp;
-	if (p->party.item != item)
+	if (p->party.item != (unsigned int)item)
 		p->party.item=item;
 
 	if (account_id == 0) {
